+++ ovp_users/helpers.py
@@ -1,11 +1,16 @@
 from django.conf import settings
+import importlib
 
 def get_settings(string="OVP_USERS"):
   return getattr(settings, string, {})
 
-def import_from_string(name):
-  components = name.split('.')
-  mod = __import__(components[0])
-  for comp in components[1:]:
-    mod = getattr(mod, comp)
-  return mod
+def import_from_string(val):
+  try:
+    # Nod to tastypie's use of importlib.
+    parts = val.split('.')
+    module_path, class_name = '.'.join(parts[:-1]), parts[-1]
+    module = importlib.import_module(module_path)
+    return getattr(module, class_name)
+  except ImportError as e:
+    msg = "Could not import '%s' for setting. %s: %s." % (val, e.__class__.__name__, e)
+    raise ImportError(msg)