+++ ovp_users/models/profile.py
@@ -4,7 +4,7 @@ from django.utils.translation import ugettext_lazy as _
 from ovp_users.helpers import get_settings, import_from_string
 
 class UserProfile(models.Model):
-  user = models.OneToOneField('User', blank=True, null=True, related_name='profile')
+  user = models.OneToOneField('User', blank=True, null=True, related_name='%(app_label)s_%(class)s_profile')
   full_name = models.CharField(_('Full name'), max_length=300, null=True, blank=True)
   skills = models.ManyToManyField('ovp_core.Skill')
   causes = models.ManyToManyField('ovp_core.Cause')